requirements.txt
Flask
Flask-SQLAlchemy
pymysql
marshmallow
cryptography

